AI-tool -- ( pdf reader)/pdf_reader/requirements.txt
pdfplumber>=0.10.0
ollama>=0.1.5
numpy>=1.24.0
python-dotenv>=1.0.0
